+++ HPL2/core/sources/graphics/PostEffect_Bloom.cpp
@@ -104,10 +104,11 @@ namespace hpl {
 	cPostEffect_Bloom::cPostEffect_Bloom(cGraphics *apGraphics, cResources *apResources, iPostEffectType *apType) : iPostEffect(apGraphics,apResources,apType)
 	{
 		cVector2l vSize = mpLowLevelGraphics->GetScreenSizeInt();
+		ePixelFormat format = cGraphics::GetHDRRendering() ? ePixelFormat_RGBA16 : ePixelFormat_RGBA;
 		
 		for(int i=0;i<2; ++i)
 		{
-			mpBlurBuffer[i] = mpGraphics->GetTempFrameBuffer(vSize/4,ePixelFormat_RGBA,i);
+			mpBlurBuffer[i] = mpGraphics->GetTempFrameBuffer(vSize/4,format,i);
 			if(mpBlurBuffer[i])
 				mpBlurTexture[i] = mpBlurBuffer[i]->GetColorBuffer(0)->ToTexture();
 		}